Refactor tests to match main code package
